pyproject.toml
[tool.poetry]
name = "switch-crew"
version = "0.1.0"
description = ""
authors = ["zhendler <zhendlerbing@gmail.com>"]
readme = "README.md"
[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.6"}
uvicorn = "^0.34.0"
pydantic-settings = "^2.7.0"
asyncpg = "^0.30.0"
sqlalchemy = {extras = ["postgresql-asyncpg"], version = "^2.0.36"}
alembic = "^1.14.0"
cloudinary = "^1.41.0"
python-multipart = "^0.0.20"
qrcode = {extras = ["pil"], version = "^8.0"}
fastapi-mail = "^1.4.2"
passlib = "^1.7.4"
bcrypt = "4.0.0"
python-jose = "^3.3.0"
pytest-cov = "^6.0.0"
sendgrid = "^6.11.0"
black = "^24.10.0"
[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.25.1"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
[[tool.poetry.packages]]
include = "src"